pyproject.toml
[project]
name = "geminibyexample-mkdocs"
version = "0.1.0"
description = "Gemini by Example converted to MkDocs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "PyYAML>=6.0",
    "pymdown-extensions>=10.0",
    "google-genai",
    "rich",
    "pydantic>=2.0.0,<3.0.0",
    "beautifulsoup4",
    "requests",
    "python-dotenv"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["geminibyexample"]

[project.optional-dependencies]
dev = [
    "ruff",
]

[tool.uv]
# UV-specific configuration

[tool.ruff]
# Assume Python 3.12
target-version = "py312"

# Same as Black.
line-length = 88

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint]
# Enable Pyflakes ('F'), pycodestyle ('E'), and import ('I') codes by default.
select = ["E", "F", "I"]
ignore = ["E501"]  # Ignore line length errors

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "I"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
